requirements.txt
requests
clicks